+++ CHANGELOG.md
@@ -17,6 +17,9 @@
 - added `publishedQuestion` resolver to `src/resolvers/versionedQuestion.ts`
 
 ### Updated
+- Updated `answer` and `addAnswerComment` permissions to include a collaborator who has `COMMENT` privileges [#243]
+- Update `removeAnswerComment` and `removeFeedbackComment` to allow deletion by creator, plan creator, or OWN-level collaborator [#243]
+- Created `commentPermissions.ts` service to determine if a comment can be deleted [#243]
 - Updated `docker-compose.yaml` to use the `dmptool-network` which will be shared with other repos
 - Updated `answer`, `fundings`, `members` and `plans` resolver to use new `hasPermissionOnPlan` function
 - Update the `tokenService` to add `dmpIds` array that stores the DMPs the user has access to